services/scout/requirements.txt
ccxt
python-dotenv
redis
